+++ packages/dev/config/eslint.cjs
@@ -100,6 +100,10 @@ module.exports = {
       { pattern: ` Copyright 20(17|18|19|20|21|22)(-${new Date().getFullYear()})? @polkadot/` },
       ' SPDX-License-Identifier: Apache-2.0'
     ], 2],
+    'import/extensions': ['error', 'ignorePackages', {
+      json: 'always',
+      jsx: 'never'
+    }],
     'import-newlines/enforce': ['error', 2048],
     'jsx-quotes': ['error', 'prefer-single'],
     'react/prop-types': [0], // this is a completely broken rule
@@ -159,7 +163,14 @@ module.exports = {
     }],
     'sort-keys': 'error',
     'spaced-comment': ['error', 'always', {
-      exceptions: ['#__PURE__']
+      block: {
+        // pure export helpers
+        markers: ['#__PURE__']
+      },
+      line: {
+        // TS reference types
+        markers: ['/ <reference']
+      }
     }]
   },
   settings: {