+++ packages/dev/scripts/polkadot-exec-node-test.mjs
@@ -26,6 +26,7 @@ const stats = {
 let logFile = null;
 let bail = false;
 let format = 'dot';
+let startAt = 0;
 
 for (let i = 0; i < args.length; i++) {
   if (args[i] === '--bail') {
@@ -43,7 +44,17 @@ for (let i = 0; i < args.length; i++) {
 
 function output (ch) {
   if (stats.total % 100 === 0) {
-    process.stdout.write('\n');
+    const now = performance.now();
+
+    if (!startAt) {
+      startAt = now;
+    }
+
+    const elapsed = (now - startAt) / 1000;
+    const m = (elapsed / 60) | 0;
+    const s = (elapsed - (m * 60));
+
+    process.stdout.write(`\n ${`${m}:${s.toFixed(3).padStart(6, '0')}`.padStart(11)}  `);
   } else if (stats.total % 10 === 0) {
     process.stdout.write('  ');
   } else if (stats.total % 5 === 0) {
@@ -85,7 +96,7 @@ function indent (count, str = '', start = '') {
   }\n`;
 }
 
-const dotParser = new TapParser({ bail }, () => {
+function parseComplete () {
   process.stdout.write('\n');
 
   let logError = '';
@@ -97,7 +108,6 @@ const dotParser = new TapParser({ bail }, () => {
 
     if (r.diag) {
       item += indent(1, [...r.fullname.split('\n'), r.name].filter((s) => !!s).join('\n'), 'x ');
-      // item += indent(2, r.name);
       item += indent(2, `${r.diag.failureType} / ${r.diag.code}`);
       item += indent(2, r.diag.error);
 
@@ -143,39 +153,7 @@ const dotParser = new TapParser({ bail }, () => {
   }
 
   process.exit(stats.fail.length);
-});
-
-dotParser
-  // Ignore the comments for now - it is mostly timing and overlaps with
-  // the actual diagnostic information
-  //
-  // .on('comment', (r) => {
-  //   stats.comm.push(r);
-  // })
-  // .on('extra', (r) => {
-  //   stats.extr.push(r);
-  // })
-  //
-  // just in-case, we want these logged
-  //
-  // .on('bailout', (r) => console.error('bailout', r))
-  // .on('plan', (r) => console.error('plan', r))
-  .on('fail', (r) => {
-    stats.fail.push(r);
-    output('x');
-  })
-  .on('pass', (r) => {
-    stats.pass.push(r);
-    output('·');
-  })
-  .on('skip', (r) => {
-    stats.skip.push(r);
-    output('>');
-  })
-  .on('todo', (r) => {
-    stats.todo.push(r);
-    output('!');
-  });
+}
 
 // 1hr default timeout ... just in-case something goes wrong on an
 // CI-like environment, don't expect this to be hit (never say never)
@@ -193,6 +171,36 @@ run({ files, timeout: 3_600_000 })
   })
   .pipe(
     format === 'dot'
-      ? dotParser
+      ? new TapParser({ bail }, parseComplete)
+      // Ignore the comments for now - it is mostly timing and overlaps with
+      // the actual diagnostic information
+      //
+      // .on('comment', (r) => {
+      //   stats.comm.push(r);
+      // })
+      // .on('extra', (r) => {
+      //   stats.extr.push(r);
+      // })
+      //
+      // just in-case, we want these logged
+      //
+      // .on('bailout', (r) => console.error('bailout', r))
+      // .on('plan', (r) => console.error('plan', r))
+        .on('fail', (r) => {
+          stats.fail.push(r);
+          output('x');
+        })
+        .on('pass', (r) => {
+          stats.pass.push(r);
+          output('·');
+        })
+        .on('skip', (r) => {
+          stats.skip.push(r);
+          output('>');
+        })
+        .on('todo', (r) => {
+          stats.todo.push(r);
+          output('!');
+        })
       : process.stdout
   );