+++ packages/dev/src/rootJs/index.ts
@@ -2,18 +2,18 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /** This should appear as-is in the output with: 1. extension added, 2. augmented.d.ts correct */
-import './augmented';
+import './augmented.js';
 
 /** This import should appear as-in in the ouput (cjs without asserts) */
 import testJson from '@polkadot/dev/rootJs/testJson.json' assert { type: 'json' };
 
 /** Double double work, i.e. re-exports */
-export { Clazz } from './Clazz';
+export { Clazz } from './Clazz.js';
 
 /** Function to ensure that BigInt does not have the Babel Math.pow() transform */
 export function bigIntExp (): bigint {
   // 123_456n * 137_858_491_849n
-  return 123_456n * (13n ** 10n);
+  return 123_456_789n * (13n ** 10n);
 }
 
 /** Function to ensure that dynamic imports work */
@@ -36,3 +36,6 @@ export function json (): string {
 export function jsOpExp (a?: number): number {
   return a ?? 42;
 }
+
+/** This is an actual check to ensure PURE is all-happy */
+export const pureOpExp = /*#__PURE__*/ jsOpExp();