+++ packages/dev/src/rootTests.ts
@@ -1,7 +1,9 @@
 // Copyright 2017-2023 @polkadot/dev authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import * as testRoot from './root';
+/// <reference types ="@polkadot/dev-test/node.d.ts" />
+
+import * as testRoot from './root.js';
 
 export function runTests ({ Clazz, TEST_PURE, bigIntExp, dynamic, jsOpExp, json }: typeof testRoot): void {
   describe('Clazz', (): void => {
@@ -34,7 +36,7 @@ export function runTests ({ Clazz, TEST_PURE, bigIntExp, dynamic, jsOpExp, json 
 
   describe('bigIntExp()', (): void => {
     it('should return the correct value', (): void => {
-      expect(bigIntExp()).toBe(123_456n * 137_858_491_849n);
+      expect(bigIntExp()).toBe(123_456_789n * 137_858_491_849n);
     });
   });