+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # CHANGELOG
 
+## 0.72.1
+
+- Split `@polkadot/dev-ts` & `@polkadot/dev-test` packages
+
+
+## 0.71.1
+
+- Ensure all `src/*` has `.js` extensions (as per ESM, eslint rules, build updates)
+
+
+## 0.70.1
+
+- Remove Babel (all compilation via tsc)
+
+
 ## 0.69.1
 
 - Remove Jest